Remove tests/context.py
This removes the context.py shim from the test suite, which was used to
ensure that tests are always run against the local directory. Instead,
it would be preferable to use `tox` (where you know the correct version
is installed).

If it is desirable to run the tests against the version in a local
directory without running multiple install steps, `pip install -e .`
will create an "editable install" that accomplishes this.