Configure tox to test against an installed package
Running tests against the package as it is installed, rather than in the
local directory, avoids some possible sources of errors, as it more
closely replicates what the end user will experience and tests that the
build and install step constructs a package that can run the tests.

This modifies the tox configuration to run the tests from a directory
other than the repository root, to ensure that the sortedcontainers/
folder is not on the Python path. An alternative (and probably better)
approach to this would be to use a src/ layout.

Ref: https://blog.ganssle.io/articles/2019/08/test-as-installed.html