Migrate to declarative build configuration
The current best practice in Python packaging is to use delcarative
build files in favor of setup.py wherever possible. This migrates
nearly all the configuration from `setup.py` to `setup.cfg`.

The only real change here is that the `name` attribute is no longer
single-sourced, but since this seems very stable that seems like a
reasonable trade-off.