Add build check to the tox configuration
This adds a tox environment that builds the project and checks the
packages for common problems with the metadata before upload.

The build artifacts will persist in .tox/build-check/tmp/dist until the
environment is run again, so this can also be used to build the project
for release if desirable.